+++ src/main/java/edu/byteme/views/menu/Frame.java
@@ -19,7 +19,6 @@ import edu.byteme.data.repositories.ClientRepository;
 import edu.byteme.data.repositories.MenuRepository;
 import edu.byteme.services.OrderService;
 import edu.byteme.views.MainLayout;
-import edu.byteme.views.orders.OrderTimeLine;
 import jakarta.annotation.security.PermitAll;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
