+++ src/main/java/edu/byteme/views/menu/OrderTimeLine.java
@@ -1,4 +1,4 @@
-package edu.byteme.views.orders;
+package edu.byteme.views.menu;
 
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.H2;
